labview-sim/requirements.txt
flask
flask_cors
flasgger==0.9.7b2
waitress
python-dotenv